requirements.txt
flask
pymysql
flask-sqlalchemy
python-dotenv
zenora
pyjwt
flask-cors
gunicorn